DocuBot/requirements.txt
faiss_cpu==1.8.0.post1
krutrim_cloud==0.6.0
numpy==2.1.1
PyPDF2==3.0.1
python-dotenv==1.0.1
sentence_transformers==3.1.1
streamlit==1.38.0


